Adds preservation_event to file_set model
