+++ app/models/preservation_workflow.rb
@@ -4,6 +4,7 @@ require 'json'
 
 class PreservationWorkflow < ActiveTriples::Resource
   include ActiveTriples::RDFSource
+  include PreservationUri
 
   property :workflow_type, predicate: "http://metadata.emory.edu/vocab/cor-terms#workflowType", multiple: false
   property :workflow_notes, predicate: "http://metadata.emory.edu/vocab/cor-terms#workflowNote", multiple: false
@@ -12,25 +13,6 @@ class PreservationWorkflow < ActiveTriples::Resource
   property :workflow_rights_basis_date, predicate: "http://metadata.emory.edu/vocab/cor-terms#workflowRightsBasisDate", multiple: false
   property :workflow_rights_basis_reviewer, predicate: "http://metadata.emory.edu/vocab/cor-terms#workflowRightsBasisReviewer", multiple: false
   property :workflow_rights_basis_uri, predicate: "http://metadata.emory.edu/vocab/cor-terms#workflowRightsBasisURI", multiple: false
-
-  # We need to convert the URI on initialize so that
-  # ActiveTriples can create a 'hash URI' for this
-  # resource.  This is necessary so that we can edit
-  # nested committee members within the ETD edit form.
-  # (Without the hash URI, we wouldn't be able to edit
-  # the committee member in the same sparql query as the
-  # ETD.)
-  # This code was taken from an example spec in the
-  # active-fedora gem:
-  # spec/integration/nested_hash_resources_spec.rb
-  def initialize(uri, parent)
-    if uri.try(:node?)
-      uri = RDF::URI("#nested_#{uri.to_s.gsub('_:', '')}")
-    elsif uri.start_with?("#")
-      uri = RDF::URI(uri)
-    end
-    super
-  end
 
   def preservation_terms
     attributes_map = { 'workflow_type' => workflow_type.first,
